Added nthreads functionality (read this commit's description)
There are server requests that, because of the limit of threads, stay in the FIFO buffer, which makes it so that the client program doesn't terminate because it has threads that are stuck. We need to find a way to, at the end of the server program, to answer those requests with 2LATE. I think that switching to alarm signals would be a good option.